+++ client/src/contexts/index.js
@@ -1,5 +1,4 @@
 import ContextProvider from './ContextProvider';
-import AnswerProvider, { useAnswer } from './AnswerContext';
 import QuestionProvider, { useQuestion } from './QuestionContext';
 import UserProvider, { useUser } from './UserContext';
 import FormProvider, { useForm } from './FormContext';
@@ -7,11 +6,9 @@ import ErrorProvider, { useError } from './ErrorContext';
 import ModeProvider, { useMode } from './ModeContext';
 
 export {
-    AnswerProvider,
     ContextProvider,
     QuestionProvider,
     UserProvider,
-    useAnswer,
     useQuestion,
     useUser,
     FormProvider,
